fix: TodoState types
