+++ src/stores/todo.ts
@@ -1,27 +1,6 @@
 import { defineStore } from "pinia";
 import { v4 as uuid } from "uuid";
-import { onUpdated } from "vue";
-
-export interface Todo {
-  id: string;
-  title: string;
-  done: boolean;
-  createAt: Date;
-  updateAt: Date;
-}
-
-export interface TodoAdd{
-  title: string;
-}
-
-export interface TodoUpdate{
-    title?: string;
-    done?: boolean;
-}
-
-export interface TodoState {
-  items: Todo[] | undefined[]  
-}
+import type {Todo, TodoAdd, TodoUpdate, TodoState} from "../types"
 
 const state = (): TodoState => ({
   items: []
@@ -33,7 +12,7 @@ const getters = {
   },
   getOrderedTodos: (state: TodoState) =>
     [...state.items].sort(
-      (a: Todo, b: Todo) => a.createdAt.getTime() - b.createdAt.getTime()     
+      (a: Todo, b: Todo) => a.createAt.getTime() - b.createAt.getTime()     
     ),
 }
 
@@ -48,17 +27,11 @@ const actions = {
     }
     this.items.push(todo)
   },
-  remove(this:any, id: string) {
+  remove(this:TodoState, id: string) {
     this.items = this.items.filter(item => item.id !== id)
   },
-  update(this:any, id: string, update: TodoUpdate) {
-    /*this.items = this.items.map(item => 
-      item.id === id
-        ? { ...item, ...update, updateAt: new Date() }
-          : item
-    )*/
-    const index = 
-    this.items.findIndex(item => item.id === id)
+  update(this:TodoState, id: string, update: TodoUpdate) {
+    const index = this.items.findIndex(item => item.id === id)
     this.items[index] = {
       ...this.items[index],
       ...update,
